Move generated classes to same package
