+++ java/merchant/pom.xml
@@ -80,7 +80,7 @@
                             <generateModelDocumentation>false</generateModelDocumentation>
                             <generateModelTests>false</generateModelTests>
                             <generateApis>false</generateApis>
-                            <modelPackage>no.bankaxept.epayment.swagger.merchant.api</modelPackage>
+                            <modelPackage>no.bankaxept.epayment.swagger.merchant</modelPackage>
                         </configuration>
                     </execution>
                 </executions>
