+++ java/merchant/src/test/java/no/bankaxept/epayment/client/merchant/MerchantClientTest.java
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.matching.EqualToPattern;
 import no.bankaxept.client.test.AbstractBaseClientWireMockTest;
-import no.bankaxept.epayment.swagger.merchant.api.Amount;
-import no.bankaxept.epayment.swagger.merchant.api.PaymentRequest;
+import no.bankaxept.epayment.swagger.merchant.Amount;
+import no.bankaxept.epayment.swagger.merchant.PaymentRequest;
 import org.junit.jupiter.api.Test;
 import reactor.adapter.JdkFlowAdapter;
 import reactor.test.StepVerifier;
